requirements.txt
Babel==1.3
Django==1.7.7
Pillow==2.4.0
Unidecode==0.04.17
argparse==1.2.1
django-appconf==1.0.1
django-compressor==1.4
django-configurations==0.8
django-extra-views==0.6.4
django-haystack==2.3.1
django-oscar==1.0.2
django-tables2==0.15.0
django-treebeard==2.0
factory-boy==2.4.1
mock==1.0.1
phonenumbers==6.3.0
psycopg2==2.6
purl==1.0.3
pycountry==1.10
pytz==2015.2
six==1.9.0
sorl-thumbnail==11.12.1b
wsgiref==0.1.2
